requirements.txt
Django==1.11.2
djangorestframework==3.8.2
pytz==2018.4
